\name{boostmtree.news}
\alias{boostmtree.news}
\title{Show the NEWS file}
\description{
  Show the NEWS file of the \pkg{boostmtree} package.
}
\usage{
boostmtree.news(...)
}
\arguments{
  \item{...}{Further arguments passed to or from other methods.}
}
\value{
  No return value, called for side effects (displays the NEWS file).
}
\author{
Amol Pande, Udaya B. Kogalur and Hemant Ishwaran
}
\examples{
if (interactive()) boostmtree.news()
}
\keyword{documentation}
