% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostmtree.news.R
\name{boostmtree.news}
\alias{boostmtree.news}
\title{Show the NEWS file}
\usage{
boostmtree.news(...)
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}
}
\value{
None.
}
\description{
Show the NEWS file of the \pkg{boostmtree} package.
}
\author{
Hemant Ishwaran, Amol Pande and Udaya B. Kogalur
}
\keyword{documentation}
