\name{boostmtree.news}
\alias{boostmtree.news}
\title{Show the NEWS file}
\description{
  Show the NEWS file of the \pkg{boostmtree} package.
}
\usage{
boostmtree.news(...)
}
\arguments{
  \item{...}{Further arguments passed to or from other methods.}
}
\value{
  None.
}
\author{
  Hemant Ishwaran, Amol Pande  and Udaya B. Kogalur
}
\keyword{documentation}
