% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.boostmtree.R
\name{print.boostmtree}
\alias{print.boostmtree}
\title{Print Summary Output}
\usage{
\method{print}{boostmtree}(x, ...)
}
\arguments{
\item{x}{An object of class \code{(boostmtree, grow)} or \code{(boostmtree,
predict)}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the \code{boostmtree} object \code{x}, following
  the convention for print methods.
}
\description{
Print summary output from the boosting analysis.
}
\references{
Pande A., Li L., Rajeswaran J., Ehrlinger J., Kogalur U.B.,
Blackstone E.H., Ishwaran H. (2017).  Boosted multivariate trees for
longitudinal data, \emph{Machine Learning}, 106(2): 277--305.
}
\author{
Hemant Ishwaran, Amol Pande and Udaya B. Kogalur
}
\keyword{print}
