\name{print.boostmtree}
\alias{print.boostmtree}
\title{Print Summary Output}
\description{
  Print summary output from the boosting analysis.
}
\usage{\method{print}{boostmtree}(x, ...)}
\arguments{
    \item{x}{An object of class \code{(boostmtree, grow)} or \code{(boostmtree,
	predict)}.}
    \item{...}{Further arguments passed to or from other methods.}
}
\author{
  Hemant Ishwaran, Pande A. and Udaya B. Kogalur
}
\references{
  Pande A., Li L., Rajeswaran J., Ehrlinger J., Kogalur U.B.,
  Blackstone E.H., Ishwaran H. (2016).  Boosted multivariate trees for
  longitudinal data.
}
\keyword{print}
